requirements.txt
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.13.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
urllib3==2.2.1
Werkzeug==3.0.3
